+++ src/CppUTest/SimpleString.cpp
@@ -141,12 +141,12 @@ void SimpleString::setStringAllocator(TestMemoryAllocator* allocator)
 }
 
 /* Avoid using the memory leak detector INSIDE SimpleString as its used inside the detector */
-char* SimpleString::allocStringBuffer(size_t _size, const char* file, int line)
+char* SimpleString::allocStringBuffer(size_t _size, const char* file, size_t line)
 {
     return getStringAllocator()->alloc_memory(_size, file, line);
 }
 
-void SimpleString::deallocStringBuffer(char* str, const char* file, int line)
+void SimpleString::deallocStringBuffer(char* str, const char* file, size_t line)
 {
     getStringAllocator()->free_memory(str, file, line);
 }
